types: fixing some type problems against other architectures

+++ source/common/buffer/buffer_impl.h
@@ -100,7 +100,7 @@ public:
     if (available_size == 0) {
       return {nullptr, 0};
     }
-    uint64_t reservation_size = std::min(size, available_size);
+    size_t reservation_size = std::min(size, available_size);
     void* reservation = &(base_[reservable_]);
     reservation_outstanding_ = true;
     return {reservation, reservation_size};
